Instrumented (but commented out) system delays
System delays on the rasp pi are unacceptably high, and more than reasonably
high on standard machines. I conclude therefore that the issue is with the rasp
pi hardware & operating support thereof, not with the codebase. Might be nice
at some point to find out the problems with the pi to resolve, but not now.