Waypoint ubuntu service created/resilience changes
Runs the waypoint code as an ubuntu service. It is now resilient against
the reader being unplugged and replugged back in. It is also resilient
against the arduino actuator not being plugged in at start up.

Installation of the debian package created using "make deb" leads to
the service being installed and started at boot time. This is resilient
against the hardware not being plugged in at boot time.

By "hardware" this should mean both arduino and rfidtag reader.