Fix two Arduino related issues
One is due to the guard being improperly tested due to lack of hardware.
(This resulted in the code beyond the guard trying to access a non-existant
object attribute)

The other is due to the Arduino Mega having a different USB serial naming
scheme from every other USB serial device I've come across.